fix: new file structure for stripe
